requirements.txt
flask
pytesseract
opencv-python
pillow
numpy
pdf2image
